(examples) Move interactions to prepareRoutes
