+++ examples/react/data-loading/src/index.js
@@ -3,16 +3,12 @@ import ReactDOM from "react-dom";
 import { createRouterComponent } from "@curi/react-dom";
 import { browser } from "@hickory/browser";
 import { createRouter } from "@curi/router";
-import prefetch from "@curi/route-prefetch";
 
 import routes from "./routes";
 import App from "./components/App";
 
-const router = createRouter(browser, routes, {
-  route: [prefetch()]
-});
+const router = createRouter(browser, routes);
 const Router = createRouterComponent(router);
-const root = document.getElementById("root");
 
 router.once(() => {
   ReactDOM.render(
